requirements.txt
Django==1.9
whitenoise==3.2.1
django-debug-toolbar==1.5
